Moved pyqtgraph import subprocess to dbviewer - pangui is already too busy to have that in there - most computers have pyqtgraph now anyways - dbviewer now depends on pyqtgraph quite a bit though - so I decided to relocate the installer instead of delete it
