+++ guis/panel/pangui/pangui.py
@@ -78,29 +78,11 @@ from PyQt5.QtWidgets import (
     QGridLayout,
 )
 
-#####################################################################
-# pyqtgraph is used by the hv gui, but none of the computers
-# have is as of 6/2/2021, so it must be installed in order to
-# import the hv gui and not crash.  This little block and 
-# installPG() can be removed after a while, but not until
-# most if not all of the computers have pyqtgraph on them
-def installPG():
-    logger.info("Attempting to install pyqtgraph...")
-    try:
-        subprocess.check_call([sys.executable, "-m", "pip", "install", "pyqtgraph", "--user"])
-    except:
-        logger.critical("Unable to install pyqtgraph, please contact a member of the software team for help.")
-        #logger.critical("Run 'pip install pyqtgraph --user' on the command line to fix this error")
-        exit()
-    logger.info("Successfully installed pyqtgraph.")
-    print("Heyyyyy it worked!")
-    import pyqtgraph
-
 try:
     import pyqtgraph
 except:
-    installPG()
-#####################################################################
+    logger.error("pyqtgraph not installed.  Run the following output line on the terminal.")
+    logger.info("pip install pyqtgraph --user")
 
 # the next import is the class for the ui
 # edit it via Qt Designer
